requirements.txt
langchain
langchain-openai
langchain-experimental
requests
beautifulsoup4
baidusearch
python-dotenv
pytest
